+++ HPL2/core/sources/graphics/Graphics.cpp
@@ -193,17 +193,17 @@ namespace hpl {
 		uint32_t selectedAdapterIdx = 0;
 		for( size_t i = 1; i < numAdapters; i++ ) {
 			if( physicalAdapters[i].type > physicalAdapters[selectedAdapterIdx].type )
-				selectedAdapterIdx = i;
+				selectedAdapterIdx = static_cast<uint32_t>(i);
 			if( physicalAdapters[i].type < physicalAdapters[selectedAdapterIdx].type )
 				continue;
 
 			if( physicalAdapters[i].presetLevel > physicalAdapters[selectedAdapterIdx].presetLevel ) 
-				selectedAdapterIdx = i;
+				selectedAdapterIdx = static_cast<uint32_t>(i);
 			if( physicalAdapters[i].presetLevel < physicalAdapters[selectedAdapterIdx].presetLevel )
 				continue;
 			
 			if(physicalAdapters[i].videoMemorySize > physicalAdapters[selectedAdapterIdx].videoMemorySize) 
-				selectedAdapterIdx = i;
+				selectedAdapterIdx = static_cast<uint32_t>(i);
 		}
 		struct RIDeviceDesc deviceInit = { 0 };
 		deviceInit.physicalAdapter = &physicalAdapters[selectedAdapterIdx];
